fix: isolate file thread mapping lookup
