+++ apps/api/src/handlers/chat/resolve-file-thread.ts
@@ -11,6 +11,10 @@ import { estimateChatContextPromptTokens } from "@/api/handlers/chat/chat-prompt
 import { computeThreadContextUsage } from "@/api/handlers/chat/compaction";
 import type { ThreadContextUsage } from "@/api/handlers/chat/compaction";
 import { resolveChatCompactionBudget } from "@/api/handlers/chat/compaction-budget";
+import {
+  lockFileChatThreadMapping,
+  type FileThreadLookupInput,
+} from "@/api/handlers/chat/file-thread-mapping";
 import { loadWindowedThreadMessages } from "@/api/handlers/chat/history-window";
 import type { ClientMessage } from "@/api/handlers/chat/message-page";
 import { loadChatMessagePage } from "@/api/handlers/chat/message-page";
@@ -57,14 +61,6 @@ const config = {
 } satisfies HandlerConfig;
 
 const CHAT_THREAD_TITLE_MAX_LENGTH = 255;
-
-type FileThreadLookupInput = {
-  entityId: SafeId<"entity">;
-  fieldId: SafeId<"field">;
-  organizationId: SafeId<"organization">;
-  userId: SafeId<"user">;
-  workspaceId: SafeId<"workspace">;
-};
 
 /**
  * Thread columns needed to load its initial message page. Fetched inline
@@ -274,45 +270,6 @@ const loadResolvedThreadMessagePage = async ({
     context,
   };
 };
-
-export const lockFileChatThreadMapping = async (
-  tx: Transaction,
-  {
-    entityId,
-    fieldId,
-    organizationId,
-    userId,
-    workspaceId,
-  }: FileThreadLookupInput,
-) =>
-  (
-    await tx
-      .select({
-        id: fileChatThreads.id,
-        mappedChatThreadId: fileChatThreads.chatThreadId,
-        thread: {
-          id: chatThreads.id,
-          chatModel: chatThreads.chatModel,
-          chatReasoningEffort: chatThreads.chatReasoningEffort,
-          contextMatterIds: chatThreads.contextMatterIds,
-          usedAnonymization: chatThreads.usedAnonymization,
-          webSearchEnabled: chatThreads.webSearchEnabled,
-        },
-      })
-      .from(fileChatThreads)
-      .leftJoin(chatThreads, eq(fileChatThreads.chatThreadId, chatThreads.id))
-      .where(
-        and(
-          eq(fileChatThreads.entityId, entityId),
-          eq(fileChatThreads.fieldId, fieldId),
-          eq(fileChatThreads.organizationId, organizationId),
-          eq(fileChatThreads.userId, userId),
-          eq(fileChatThreads.workspaceId, workspaceId),
-        ),
-      )
-      .limit(1)
-      .for("update", { of: fileChatThreads })
-  ).at(0);
 
 const findFieldKeyedChatThread = async (
   tx: Transaction,
